fix: address #621 review findings
- in_memory transport: requeue a copy on ack_wait expiry and nack so stale
  delivery handles become inert no-ops (SQS receipt-handle parity), preserving
  one-in-flight-per-group; regression tests added
- request handler SSE: deterministically close the store chunk generator
  (releases per-request state) and stop surfacing internal exception details
  in error frames (CodeQL); timeouts keep their safe store-authored text
- friendlier local defaults for long agent runs: in_memory ack_wait default
  300s, default response wait budget 60x1s when no response_store block;
  prominent caution added to the queue-mode guide's local section
- config: no_of_consumers descriptions now cover the in-process pipeline;
  _get_retry_config annotation corrected to tuple[int, float]; em dash
  removed from the _run_async_sync docstring
- spec/design back-edits: pipeline-imports-core-and-api coupling and the
  ChatService-wrapper runner layer recorded in design.md; spec name drift
  fixed (in_memory field, IN_MEMORY enum, testing.py in layout); spec §12.1
  now enumerates the bounded-wait behavioral deltas

+++ ak-py/src/agentkernel/pipeline/response_store/base.py
@@ -74,7 +74,7 @@ class ResponseStore(ABC):
         return None
 
     @staticmethod
-    def _get_retry_config() -> tuple[int, int]:
+    def _get_retry_config() -> tuple[int, float]:
         """Read (retry_count, delay) for get_message_with_retry from config."""
         response_store_config = AKConfig.get().execution.response_store
         return response_store_config.retry_count, response_store_config.delay
